% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfrog.R
\name{leapfrog.grad}
\alias{leapfrog.grad}
\title{Leapfrog Integrator for Hamiltonian Dynamics given the gradient}
\usage{
leapfrog.grad(g, q, p, L, eps)
}
\arguments{
\item{g}{Gradient function for the system of interest.}

\item{q}{Initial position of the system.}

\item{p}{Initial momentum of the system.}

\item{L}{Number of leapfrog steps.}

\item{eps}{Size of each leapfrog step.}
}
\value{
A list containing both the position `q` and momentum `p` at the end
        of the integrator.
}
\description{
Performs leapfrog steps for the Hamiltonian dynamics system given the
gradient function.
}
