% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfrog.R
\name{leapfrog}
\alias{leapfrog}
\title{Leapfrog Integrator for Hamiltonian Dynamics}
\usage{
leapfrog(f, q, p, L, eps)
}
\arguments{
\item{f}{Function of the system of interest.}

\item{q}{Initial position of the system.}

\item{p}{Initial momentum of the system.}

\item{L}{Number of leapfrog steps.}

\item{eps}{Size of each leapfrog step.}
}
\value{
A list containing both the position `q` and momentum `p` at the end
        of the integrator.
}
\description{
Performs leapfrog steps for the Hamiltonian dynamics system given the
integral of the gradient function for the system of interest, using
numerical derivatives for the gradient.
}
