% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfrog.R
\name{num.grad}
\alias{num.grad}
\title{Numerical Approximation of the Gradient}
\usage{
num.grad(f, x)
}
\arguments{
\item{f}{Function to differentiate.}

\item{x}{Point to calculate differential.}
}
\value{
Approximation of the gradient.
}
\description{
Performs simple numerical differentiation at a specific point.
}
